refactor: move create-gtag inside core
