+++ src/core/create-gtag/index.ts
@@ -1,0 +1,1 @@
+export * from "./create-gtag";
