+++ src/core/index.ts
@@ -1,0 +1,2 @@
+export * from "./add-gtag";
+export * from "./create-gtag";
